requirements.txt
absl-py==1.4.0
alphafold-colabfold==2.3.4
appdirs==1.4.4
astunparse==1.6.3
biopython==1.79
blessed==1.20.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
chex==0.1.6
click==8.1.3
colabfold @ git+https://github.com/sokrypton/ColabFold@05c0cb38d002180da3b58cdc53ea45a6b2a62d31
conda-pack==0.7.1
contextlib2==21.6.0
contourpy==1.0.7
cycler==0.11.0
dm-haiku==0.0.9
dm-tree==0.1.8
docker==6.0.1
et-xmlfile @ file:///home/conda/feedstock_root/build_artifacts/et_xmlfile_1674664118162/work
filelock==3.16.1
Flask==2.2.3
flatbuffers==23.3.3
fonttools==4.39.3
fsspec==2024.10.0
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpustat==1.1.1
grpcio==1.53.0
h5py==3.8.0
huggingface-hub==0.26.1
idna==3.4
immutabledict==2.2.3
importlib-metadata==4.13.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jax==0.3.25
jaxlib @ https://storage.googleapis.com/jax-releases/cuda11/jaxlib-0.3.25+cuda11.cudnn82-cp39-cp39-manylinux2014_x86_64.whl
Jinja2==3.1.2
jmp==0.0.4
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-collections==0.1.1
numpy==1.23.5
nvidia-ml-py==12.535.108
oauthlib==3.2.2
openpyxl @ file:///home/conda/feedstock_root/build_artifacts/openpyxl_1723459101873/work
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==4.22.1
psutil==5.9.6
py3Dmol==2.0.1.post1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0
regex==2024.9.11
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.5
scipy==1.10.1
six==1.16.0
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
tokenizers==0.20.1
toolz==0.12.0
torch==1.13.1+cu117
torchaudio==0.13.1
torchvision==0.14.1+cu117
tqdm==4.65.0
transformers==4.46.0
typing_extensions==4.5.0
tzdata==2024.2
urllib3==1.26.15
wcwidth==0.2.8
websocket-client==1.5.1
Werkzeug==2.2.3
wrapt==1.14.1
WTForms==3.0.1
zipp==3.15.0
